Compound Interest Calculator
As of now the calculator will give you the amount of interest garnered over the specified time period. It will also provide you with the total worth of your investment. 

One thing that I need to add later is a decimal format for the results. Right now it just spits out a number with 10+ decimal points attached to it.